Pull Viewport into a module
Fixes #253
